refactor: 컨트롤러 이름변경 & 객체 분리
CarsController -> Game으로 변경
Game에서 round마다 경기하는 로직을 Race 클래스로 분리함
